use $SUDOCMD to run vm commands - add user SetupComplete as last step
One of the problems with make-ad-vm.sh is that there was no way to know
when it was complete.  setupscript99.cmd will create an AD user called
SetupComplete as the very last step.  When this user is present, the
setup process is finished.  The make-ad-vm.sh script will delete this
user entry.
Some of the virsh and vm commands can only be run by root.  These commands
are now prefixed with $SUDOCMD.  If you are running the script as a non-root
user, and you have sudo setup to allow that user to run the commands, you
can use
 SUDOCMD=sudo ..othervars.. make-ad-vm.sh ....
Otherwise, if you are running as root, just leave SUDOCMD blank.