support images/backing images, conf files, vm completion
The biggest change is support for VM images, including using
a backing image.
This commit also supports passing in
config files on the command line rather than having to specify
all of the parameters as environment variables.
You can use a different virtual network other than default - the
VM_NETWORK_NAME parameter.
The old method of polling for an LDAP entry using the CA cert is
not suitable for the general case.  Instead, you now specify a
VM_WAIT_FILE (using DOS backslashes \\ in the file name).  The
script will use virt-cat to test if the file exists.  The last
setupscript99.cmd will create the file at the very end of setup.